moving validators to BLL
