+++ src/EShop.BLL/EShop.BLL.csproj
@@ -6,8 +6,9 @@
         <Nullable>enable</Nullable>
     </PropertyGroup>
 
-    <ItemGroup>
-      <ProjectReference Include="..\EShop.DAL\EShop.DAL.csproj" />
+	<ItemGroup>
+		<PackageReference Include="FluentValidation" Version="11.9.2" />
+		<ProjectReference Include="..\EShop.DAL\EShop.DAL.csproj" />
     </ItemGroup>
 
 </Project>
