+++ src/EShop.BLL/Validators/OrderItemValidator.cs
@@ -1,0 +1,11 @@
+﻿using EShop.DAL.Entities;
+using FluentValidation;
+
+namespace EShop.BLL.Validators;
+
+public class OrderItemValidator: AbstractValidator<OrderItem>
+{
+    public OrderItemValidator()
+    {         
+    }
+}
