+++ src/EShop.BLL/Validators/ProductValidator.cs
@@ -1,0 +1,17 @@
+﻿using EShop.DAL.Entities;
+using FluentValidation;
+
+namespace EShop.BLL.Validators;
+
+public class ProductValidator : AbstractValidator<Product>
+{
+    public ProductValidator()
+    {
+        RuleFor(product => product.Name)
+          .NotEmpty()
+          .WithMessage("Name is empty.");
+        RuleFor(product => product.Description)
+          .NotEmpty()
+          .WithMessage("Description is empty.");
+    }
+}
