--- src/EShop.WebAPI/Validators/OrderItemValidator.cs
@@ -1,12 +1,0 @@
-﻿using EShop.DAL.Entities;
-using FluentValidation;
-
-namespace EShop.WebAPI.Validators
-{
-    public class OrderItemValidator: AbstractValidator<OrderItem>
-    {
-        public OrderItemValidator()
-        {         
-        }
-    }
-}
