--- src/EShop.WebAPI/Validators/ProductValidator.cs
@@ -1,18 +1,0 @@
-﻿using EShop.DAL.Entities;
-using FluentValidation;
-
-namespace EShop.WebAPI.Validators
-{
-    public class ProductValidator : AbstractValidator<Product>
-    {
-        public ProductValidator()
-        {
-            RuleFor(product => product.Name)
-              .NotEmpty()
-              .WithMessage("Name is empty.");
-            RuleFor(product => product.Description)
-              .NotEmpty()
-              .WithMessage("Description is empty.");
-        }
-    }
-}
